requirements.txt
django
djangorestframework
djangorestframework_simplejwt
drf-spectacular
python-dotenv
kavenegar
